requirements.txt
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
openai
bs4
faiss-cpu
python-dotenv
tqdm
tiktoken
langchain_chroma
langchainhub